[ahk] [Loot] Fix the SGL bug in TransferLootFromInventoryIntoCube()
See P run log for the description of the bug.

Fixing the bug by detecting if the loot is on the mouse. This is done by clearing the mouse to a known position (top right corner) and then detect if the mouse pixel is the mouse or something else (the loot).